+++ src/main/java/appeng/client/gui/implementations/GuiPatternValueAmount.java
@@ -24,7 +24,7 @@ import appeng.parts.reporting.PartPatternTerminalEx;
 public class GuiPatternValueAmount extends GuiAmount {
 
     private final int valueIndex;
-    private final int originalAmount;
+    private final long originalAmount;
 
     @Reflected
     public GuiPatternValueAmount(final InventoryPlayer inventoryPlayer, final ITerminalHost te) {
